Split test fixtures into separate file
